+++ app/services/renotification.rb
@@ -1,7 +1,7 @@
 class Renotification
   def self.execute
     programs = Program.active
-    WriteLog.info self, "Was executed at #{Time.now} for #{programs.size} program(s)."
+    Zazo::Tools::Logger.info(self, "executed for #{programs.size} program(s)")
     programs.each { |program| program.execute }
   end
 end